segregated files for db, connections, map_views
