+++ covid19_visualisation.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-# In[ ]:
+# In[1]:
 
 
 #Completed
@@ -15,7 +15,7 @@
 # Integrating code with matching algorithm & writing matched entry into spreadsheet
 
 
-# In[1]:
+# In[2]:
 
 
 #Basic
@@ -33,35 +33,17 @@ from googleapiclient.discovery import build
 from google_auth_oauthlib.flow import InstalledAppFlow
 from google.auth.transport.requests import Request
 
-#Map
-import keplergl
-
-#DB Connections
-import mysql.connector as sql_con
-from sqlalchemy import create_engine
-import cred_config as cc
-
 #File Transfer
 import ftplib
 import os
 
-from bs4 import BeautifulSoup
+from connections import connections
+from map_generator import public_map,private_map
+from settings import *
 
 
-# In[2]:
+# In[3]:
 
-
-
-default_r=0.5
-
-#Approximation
-lat_deg_to_km = 95.0
-lon_deg_to_km = 110.0
-buffer_radius = 1/np.sqrt(95*95+110*110)
-
-
-# If modifying these scopes, delete the file token.pickle.
-SCOPES = ['https://www.googleapis.com/auth/spreadsheets']
 
 # The ID and range of a sample spreadsheet.
 volunteer_sheet_data = [{'source':'GreenDream','sheet_id':'1e9H5yO1COGLNfA3lyZxRSgc2llDKRSFZX92Ov8VOzOs','range':'Form Responses 1!A1:K1000'}]
@@ -70,39 +52,14 @@ senior_citizen_sheet_data = [{'source':'GreenDream','sheet_id':'1KrZCG_fYvImIy_-
 public_file_name= 'output/COVID_SOS_v0.html'
 private_file_name= 'output/private_COVID_SOS_v0.html'
 
-#FORMAT: from map_config.filename import *
-from map_config.map_config_private import *
-from map_config.map_config_public import *
+
+# In[ ]:
 
 
-# In[3]:
 
-
-def map_config_fn(map_config_file):
-    with open(map_config_file,'r') as config_file_reader:
-        config_file = config_file_reader.read()
-        exec(config_file, None, locals())
-    dx = live_config.copy()
-    return dx
 
 
 # In[4]:
-
-
-def connections(con_name):
-    if(con_name=='db_read'):
-        cred_r=cc.credentials['covid_sos_read']
-        server_con = sql_con.connect(user=cred_r['user'], password=cred_r['password'], host=cred_r['host'],database=cred_r['database'])
-    if(con_name=='db_write'):
-        cred_w = cc.credentials['covid_sos_write']
-        server_con = create_engine("mysql+pymysql://{user}:{password}@{host}/{database}".format(user = cred_w['user'], password = cred_w['password'], host = cred_w['host'], database = cred_w['database']), pool_size=10, max_overflow=20, echo=False)
-    if(con_name=='ftp'):
-        FTP_con = cc.credentials['ftp']
-        server_con = ftplib.FTP(host=FTP_con['host'], user=FTP_con['user'], passwd=FTP_con['password'])
-    return server_con
-
-
-# In[5]:
 
 
 def google_api_activation():
@@ -164,51 +121,7 @@ def sheet_clean_up(df,default_r,buffer_radius):
     return f_df
 
 
-def html_file_changes(output_file_name):
-    with open(output_file_name,'r') as file:
-        filedata = file.read()
-        filedata = filedata.replace('keplergl-jupyter-html','covid-sos-page')
-        filedata = filedata.replace('UA-64694404-19','UA-143016880-1')
-    with open(output_file_name,'w') as file:
-        file.write(filedata)
-    with open(output_file_name,'r') as output_file_reader:
-        bs = output_file_reader.read()
-    soup = BeautifulSoup(bs, 'html.parser')
-    soup.title.string='COVID SOS Initiative - Connecting Volunteers with Requests'
-    #Enabling GA
-    with open(output_file_name, "w") as file:
-        file.write(str(soup))
-    return None
-
-
-# In[6]:
-
-
-def public_map(v_df,r_df,output_file_name):
-    v_df['WhatsApp Contact Number']=9582148040
-    r_df['Mobile Number']=9582148040
-    map_1 = keplergl.KeplerGl(height=800,data={'volunteer_data':v_df.loc[v_df['Lat']!=0,['Timestamp', 'Full Name','geometry','Lat','Lon','radius','icon','TYPE']],
-                                               'requests_data':r_df.loc[r_df['Lat']!=0,['Timestamp', 'Full Name', 'Mobile Number', 'Age'
-    ,'Would you like to give any special instructions to the volunteer aligned to you? Please share below.','Task Status','geometry','Lat','Lon','radius','icon','TYPE']]})
-    print('The public map contains ', v_df[v_df['Lat']!=0].shape[0],' volunteers and ', r_df[r_df['Lat']!=0].shape[0], ' pending requests')
-    #variable live_config is defined when "file" is executed
-    map_1.config = public_live_config
-    map_1.save_to_html(file_name=output_file_name)
-    html_file_changes(output_file_name)
-    push_file_to_server(output_file_name,output_file_name)
-    push_file_to_server(output_file_name,'output/share_and_survive_v0_dark.html')
-    return map_1
-
-def private_map(v_df,r_df,output_file_name):
-    r_df = r_df[r_df['Task Status']=='Pending']
-    map_1 = keplergl.KeplerGl(height=800,data={'volunteer_data':v_df[v_df['Lat']!=0],'requests_data':r_df[r_df['Lat']!=0]})
-    print('The private Map contains ', v_df[v_df['Lat']!=0].shape[0],' volunteers and ', r_df[r_df['Lat']!=0].shape[0], ' pending requests')
-    #variable live_config is defined when "file" is executed
-    map_1.config = private_live_config
-    map_1.save_to_html(file_name=output_file_name)
-    html_file_changes(output_file_name)
-    push_file_to_server(output_file_name,output_file_name)
-    return map_1
+# In[5]:
 
 
 def push_file_to_server(File2Send,Url2Store):
@@ -221,7 +134,7 @@ def push_file_to_server(File2Send,Url2Store):
     return None
 
 
-# In[7]:
+# In[6]:
 
 
 def main():
@@ -238,14 +151,17 @@ def main():
     
     r_df = sheet_clean_up(requests_df,default_r,buffer_radius)
     r_df['icon']='home'
-    
-    private_map_v1 = private_map(v_df,r_df,private_file_name)
-    public_map_v1 = public_map(v_df,r_df,public_file_name)
+
+    private_map_v1 = private_map(v_df,r_df,private_file_name,map_pkg='kepler')
+    push_file_to_server(private_file_name,private_file_name)
+    push_file_to_server(private_file_name,'output/share_and_survive_v0_dark.html')
+    public_map_v1 = public_map(v_df,r_df,public_file_name,map_pkg='kepler')
+    push_file_to_server(public_file_name,public_file_name)
     #Processing Data
     return v_df, r_df, private_map_v1,public_map_v1
 
 
-# In[8]:
+# In[7]:
 
 
 v_df, r_df, p1,p2=main()
@@ -257,7 +173,7 @@ v_df, r_df, p1,p2=main()
 
 
 
-# In[ ]:
+# In[8]:
 
 
 # with open('map_config/map_config_public.py','w') as f:
@@ -266,7 +182,7 @@ v_df, r_df, p1,p2=main()
 #     print(f.read())
 
 
-# In[ ]:
+# In[9]:
 
 
 #v_query = ("""Select * from volunteers""")
@@ -275,7 +191,7 @@ v_df, r_df, p1,p2=main()
 #v_df.to_sql(name = 'volunteers', con = engine, schema='thebang7_COVID_SOS', if_exists='append', index = False,index_label=None)
 
 
-# In[ ]:
+# In[10]:
 
 
 
@@ -289,7 +205,7 @@ v_df, r_df, p1,p2=main()
 
 
 
-# In[ ]:
+# In[11]:
 
 
 # from folium import Map, Marker, GeoJson
